requirements.txt
allure-pytest==2.14.0
allure-python-commons==2.14.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
execnet==2.1.1
Faker==37.3.0
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
prettytable==3.16.0
pydantic==2.11.5
pydantic_core==2.33.2
pytest==8.3.5
pytest-check==2.5.3
pytest-xdist==3.7.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
testit-adapter-pytest==3.6.2
testit-api-client==6.1.1
testit-python-commons==3.6.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
